Image Cache Improved default size to avoid loop dropping cached tiles in the current screen.
The default Cache Size (30Mb) is not enough for hidpi screens where the full screen map requires more
space than that cached. In this condition, when drawing the full screen new tiles are loaded that
drops the old ones, that are needed to display the screen so old tiles are reloaded dropping the new
tiles and so on.
To avoid this, the default cache size is set to 3 times the space needed to cache the whole screen.
The screen size is calculated taking into account the HiDPI multiplication -- there's no function to
access the physical screen size in Qt5.
This fix the issue.
The size can still be changed with the old set function.

Signed-off-by: Federico Fuga <fuga@studiofuga.com>